% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesToIntegers.R
\name{classesToIntegers}
\alias{classesToIntegers}
\title{Create a vector of integer classes from a vector of character classes.}
\usage{
classesToIntegers(classes)
}
\arguments{
\item{\code{classes}}{a character string of landscape classes (e.g. "Grass", "Trees", "Water", "Road", "Building").}
}
\value{
The function returns a character vector of the same length as \code{classes}, but where each unique string has been replaced by an integer (starting from 1).
}
\description{
\code{createTrainingSetRGB} creates a vector of integer classes from a vector of character classes.
}
